Update Makefile with more root agda files.
When trying to install the library with NixOS,
I noticed that some agdai files were not prebuilt by the running make.
The files could not be built on the fly and saved back into the directory
that contains the library files because the nix store is read-only.

Files I left out because of compile errors:
* string-format-issue.agda
* type-level.agda